fix(core): keep RuntimeAdapterProvider jsx out of the store entry graph
